Reorganize habit and schedule retrieval code for better readability
