+++ core/database/src/main/java/com/rendox/routinetracker/core/database/di/HabitLocalDataModule.kt
@@ -1,13 +1,13 @@
 package com.rendox.routinetracker.core.database.di
 
-import com.rendox.routinetracker.core.database.habit.DueDateLocalDataSource
-import com.rendox.routinetracker.core.database.habit.DueDateLocalDataSourceImpl
+import com.rendox.routinetracker.core.database.habit.due_dates.DueDateLocalDataSource
+import com.rendox.routinetracker.core.database.habit.due_dates.DueDateLocalDataSourceImpl
 import com.rendox.routinetracker.core.database.habit.HabitLocalDataSource
 import com.rendox.routinetracker.core.database.habit.HabitLocalDataSourceImpl
-import com.rendox.routinetracker.core.database.habit.ScheduleLocalDataSource
-import com.rendox.routinetracker.core.database.habit.ScheduleLocalDataSourceImpl
+import com.rendox.routinetracker.core.database.habit.schedule.ScheduleLocalDataSourceImpl
 import com.rendox.routinetracker.core.database.habit.WeekDaysMonthRelatedLocalDataSource
 import com.rendox.routinetracker.core.database.habit.WeekDaysMonthRelatedLocalDataSourceImpl
+import com.rendox.routinetracker.core.database.habit.schedule.ScheduleLocalDataSource
 import org.koin.core.qualifier.named
 import org.koin.dsl.module
