+++ core/database/src/main/java/com/rendox/routinetracker/core/database/habit/schedule/ScheduleLocalDataSource.kt
@@ -1,10 +1,9 @@
-package com.rendox.routinetracker.core.database.habit
+package com.rendox.routinetracker.core.database.habit.schedule
 
 import com.rendox.routinetracker.core.model.Schedule
 
 interface ScheduleLocalDataSource {
     fun insertSchedule(schedule: Schedule)
-    fun getScheduleById(scheduleId: Long): Schedule
-    fun getAllSchedules(): List<Schedule>
+    fun getScheduleById(habitId: Long): Schedule
     fun deleteSchedule(scheduleId: Long)
 }
