+++ core/database/src/main/java/com/rendox/routinetracker/core/database/model/habit/HabitType.kt
@@ -1,0 +1,5 @@
+package com.rendox.routinetracker.core.database.model.habit
+
+enum class HabitType {
+    YesNoHabit,
+}
