+++ core/database/src/main/java/com/rendox/routinetracker/core/database/model/schedule/ScheduleType.kt
@@ -1,0 +1,13 @@
+package com.rendox.routinetracker.core.database.model.schedule
+
+enum class ScheduleType {
+    EveryDaySchedule,
+    WeeklyScheduleByDueDaysOfWeek,
+    WeeklyScheduleByNumOfDueDays,
+    MonthlyScheduleByDueDatesIndices,
+    MonthlyScheduleByNumOfDueDays,
+    AlternateDaysSchedule,
+    CustomDateSchedule,
+    AnnualScheduleByDueDates,
+    AnnualScheduleByNumOfDueDays,
+}
