+++ core/database/src/main/sqldelight/com/rendox/routinetracker/core/database/habit/habitEntity.sq
@@ -1,4 +1,4 @@
-import com.rendox.routinetracker.core.database.habit.model.HabitType;
+import com.rendox.routinetracker.core.database.model.habit.HabitType;
 import kotlin.Float;
 import kotlin.Int;
